+++ packages/lib/src/index.ts
@@ -1,2 +1,4 @@
 export const isoNow = () => new Date().toISOString();
 export * from "./flags.data";
+export * from "./random.util";
+export * from "./history.store";